+++ apps/nameguard.io/src/app/components/atoms/ListSectionBadge.tsx
