+++ apps/nameguard.io/src/app/components/atoms/MobileSectionDivider.tsx
