+++ apps/nameguard.io/src/app/components/atoms/index.ts
@@ -1,2 +1,3 @@
 export * from "./icons";
 export * from "./CalButton";
+export * from "./ListSectionBadge";
