+++ apps/nameguard.io/src/app/components/molecules/NGReport.tsx
