+++ apps/nameguard.io/src/app/components/organisms/Footer.tsx
