+++ apps/nameguard.io/src/app/components/organisms/HeroCarousel.tsx
