+++ apps/nameguard.io/src/app/components/organisms/index.ts
@@ -1,0 +1,12 @@
+export * from "./ChecksSection";
+export * from "./ComingSoonSection";
+export * from "./DevelopersSection";
+export * from "./ExploreTheEcosystemSection";
+export * from "./Footer";
+export * from "./Header";
+export * from "./Hero";
+export * from "./HeroCarousel";
+export * from "./NewExitSection";
+export * from "./NotFound";
+export * from "./ReadySection";
+export * from "./RoadMap";
