+++ apps/nameguard.io/src/app/layout.tsx
@@ -5,10 +5,10 @@ import { Metadata as NamehashMetadata } from "@namehash/seo";
 import { Inter } from "next/font/google";
 import Favicon from "../../public/favicon/favicon_package_v0-2/favicon.ico";
 import AppleTouchIcon from "../../public/favicon/favicon_package_v0-2/apple-touch-icon.png";
-import { NameGuardWrapper } from "./components/NameGuardWrapper";
+import { NameGuardWrapper } from "./components/molecules/NameGuardWrapper";
 
-import { Footer } from "@/app/components/Footer";
-import { Header } from "@/app/components/Header";
+import { Footer } from "@/app/components/organisms/Footer";
+import { Header } from "@/app/components/organisms/Header";
 import React from "react";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
