+++ apps/nameguard.io/src/app/not-found.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import NotFound from "./components/NotFound";
+import NotFound from "./components/organisms/NotFound";
 
 export const metadata: Metadata = {
   title: "Not Found",
