+++ apps/nameguard.io/src/types/listSectionTypes.ts
@@ -1,0 +1,6 @@
+export type ListSectionElement = {
+  header: React.ReactNode;
+  text: React.ReactNode;
+  icon: React.ReactNode;
+  link?: string;
+};
